requirements.txt
gdown==5.2.0
fastapi==0.111.0
uvicorn==0.20.0
numpy==1.26.3
transformers==4.40.1
python-multipart==0.0.9
easyocr==1.7.1
pillow==10.2.0